+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this theme will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-07-24
+
+### Added
+
+- Added Mermaid and Mathjax equation support
+- Added posts/mermaid-mathjax/index.md
+- Added CHANGELOG.md file
+
+### Changed
+
+-Nothong
+
+### Fixed
+
+- Search function updated. Now works well in navbar and search page
+
+
 ## [1.1.0] - 2025-07-10
 
 ### Added